initialize only once

+++ app/webapp/ruby/app.rb
@@ -37,10 +37,12 @@ class App < Sinatra::Base
     db.query("DELETE FROM user WHERE id > 1000")
     db.query("DELETE FROM image WHERE id > 1001")
 
-    rows = db.query("SELECT * FROM image WHERE id <= 1001").to_a
-    rows.each do |row|
-      file_path = "/home/isucon/isubata/webapp/public/icons/#{row['name']}"
-      File.write(file_path, row['data'])
+    unless File.exists?('/home/isucon/isubata/webapp/public/icons/default.png') do
+      rows = db.query("SELECT * FROM image WHERE id <= 1001").to_a
+      rows.each do |row|
+        file_path = "/home/isucon/isubata/webapp/public/icons/#{row['name']}"
+        File.write(file_path, row['data'])
+      end
     end
 
     db.query("DELETE FROM channel WHERE id > 10")
